% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-attributes.R
\name{edges}
\alias{edges}
\alias{edges,obj3d-method}
\alias{edges,facelayer-method}
\title{The edges of a 3d object}
\usage{
edges(x)

\S4method{edges}{obj3d}(x)

\S4method{edges}{facelayer}(x)
}
\arguments{
\item{x}{(\code{\link{trigrid}}, \code{\link{hexagrid}} or \code{\link{facelayer}}) The grid or linked data object.}
}
\value{
The edges of the grid, as a \code{character} matrix.
}
\description{
Shorthand function to get the edges slot of an icosahedral grid or a grid linked to a facelayer.
}
