% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gridlayer-subset.R
\docType{methods}
\name{[,gridlayer,ANY,missing-method}
\alias{[,gridlayer,ANY,missing-method}
\alias{[,gridlayer,Extent,missing-method}
\alias{[<-,gridlayer,ANY,ANY-method}
\alias{[<-,gridlayer-method}
\title{Extraction from a gridlayer using indices}
\usage{
\S4method{[}{gridlayer,ANY,missing}(x, i)

\S4method{[}{gridlayer,Extent,missing}(x, i)

\S4method{[}{gridlayer,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{(\code{\link{facelayer}}) The object to be subsetted.}

\item{i}{(\code{logical}, \code{numeric} nor \code{extent}) The subscript vector, or extent, specifying the faces that are used for subsetting. As in \code{\link[base]{subset}}.}

\item{value}{The replacement values.}
}
\value{
The extraction methods return \code{\link{facelayer}}-class objects.
}
\description{
Shorthand to the \code{\link[icosa]{subset}} function.

Function to replace specific elements in a gridlayer object
}
\details{
All these methods are implementing direct replacement in the \code{@values} slot of a layer, depending on criteria used for subsetting.
}
