% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-attributes.R
\name{faces}
\alias{faces}
\alias{faces,trigrid-method}
\alias{faces,gridlayer-method}
\title{The faces of a 3d object}
\usage{
faces(x)

\S4method{faces}{trigrid}(x)

\S4method{faces}{gridlayer}(x)
}
\arguments{
\item{x}{(\code{\link{trigrid}}, \code{\link{hexagrid}} or \code{\link{facelayer}}) The grid or facelayer object.}
}
\value{
The faces of the grid as a \code{character} matrix.
}
\description{
Shorthand function to get the faces slot of an icosahedral grid or a grid linked to a \code{\link{facelayer}}.
}
