% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-2d-grid.R
\name{gridlabs}
\alias{gridlabs}
\title{Labels of grid vertices, faces and edges.}
\usage{
gridlabs(gridObj, type = "f", projargs = NULL, ...)
}
\arguments{
\item{gridObj}{(\code{\link{trigrid}}, \code{\link{hexagrid}}) An icosahedral grid.}

\item{type}{(\code{character}) The type of element to be plotted: either \code{"f"} (faces), \code{"v"} (vertices) or  \code{"e"} (edges).}

\item{projargs}{(\code{character} or \code{\link[sp:CRS-class]{CRS}}) A projection string for the transformation fo coordinates.}

\item{...}{Arguments passed to the \code{\link[graphics]{text}} function.}
}
\value{
The function has no return value.
}
\description{
This function will show where the grid elements are located.
}
\examples{
gr <- hexagrid(sp=TRUE)
plot(gr)
gridlabs(gr)
}
