% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{icosa}
\alias{icosa}
\title{Global Triangular and Hexa-Pentagonal Grids Based on Tessellated Icosahedra}
\description{
The \bold{icosa} package provides tools to aggregate and analyze geographic data
using grids based on tessellated icosahedra. The procedures can be set to provide a grid with a
   custom resolution. Both the primary triangular and their inverted penta-
  hexagonal grids are available for implementation. Additional functions
   are provided to position points (latitude-longitude data) on the grids,
   to allow 2D and 3D plotting, use raster data and shapefiles.
}
\details{
This is still the Beta version. Notes about found bugs and suggestions are more than welcome!
}
\examples{
# Create a triangular grid
tri <- trigrid(c(2,2))
}
\author{
Adam T. Kocsis (adam.t.kocsis@gmail.com)
}
