% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-attributes.R,
%   R/data-gridlayer-attributes.R
\name{length,trigrid-method}
\alias{length,trigrid-method}
\alias{length,gridlayer-method}
\title{The number of faces in a \code{trigrid} or \code{hexagrid} class object.}
\usage{
\S4method{length}{trigrid}(x)

\S4method{length}{gridlayer}(x)
}
\arguments{
\item{x}{(\code{\link{trigrid}}, \code{\link{hexagrid}} or \code{\link{facelayer}}) The object.}
}
\value{
An integer value.
}
\description{
The length of the object is interpreted as the number of faces it contains.
}
