% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-2d-grid.R
\name{lines,trigrid-method}
\alias{lines,trigrid-method}
\title{Lines method for the \code{trigrid} and \code{hexagrid} classes}
\usage{
\S4method{lines}{trigrid}(x, projargs = NULL, ...)
}
\arguments{
\item{x}{(\code{\link{trigrid}}, \code{\link{hexagrid}}) Object.}

\item{projargs}{(\code{character} or \code{\link[sp:CRS-class]{CRS}}) A projection string for the transformation fo coordinates.}

\item{...}{Arguments passed to the \code{\link[sp:panel]{sp.lines}} method.}
}
\value{
The function has no return value.
}
\description{
This function will invoke the \code{\link[sp:panel]{sp.lines}} method of the \code{\link[sp]{SpatialPolygons}} class.
}
