% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-sp-polygons.R
\name{newsp}
\alias{newsp}
\alias{newsp,trigrid-method}
\title{Add a \code{\link[sp]{SpatialPolygons}} object to a predefined slot in a \code{\link{trigrid}} or \code{\link{hexagrid}} object}
\usage{
newsp(gridObj, res = NULL)

\S4method{newsp}{trigrid}(gridObj, res = NULL)
}
\arguments{
\item{gridObj}{(\code{\link{trigrid}} or \code{\link{hexagrid}}) An icosahedral grid.}

\item{res}{(\code{numeric}) The number of points inserted between two vertices, passed to \code{\link{SpPolygons}}.}
}
\value{
A \code{\link{trigrid}} or \code{\link{hexagrid}} object with the new \code{@sp} slot.
}
\description{
Add a \code{\link[sp]{SpatialPolygons}} object to a predefined slot in a \code{\link{trigrid}} or \code{\link{hexagrid}} object
}
\examples{
a<-trigrid(4)
a<-newsp(a)
plot(a)
}
