% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\docType{methods}
\name{orientation}
\alias{orientation}
\alias{orientation,}
\alias{orientation,trigrid-method}
\alias{orientation-method}
\alias{orientation<-}
\alias{orientation<-,trigrid-method}
\alias{trigrid-orientation-method}
\alias{trigrid-setorientation-method}
\title{Extracting the grid orientation}
\usage{
orientation(gridObj, ...)

\S4method{orientation}{trigrid}(gridObj, display = "deg", ...)

orientation(gridObj) <- value

\S4method{orientation}{trigrid}(gridObj) <- value
}
\arguments{
\item{gridObj}{The grid object.}

\item{...}{values passed on to the rotate() function.}

\item{display}{The output units. In case it is set to "deg" the output will be in degrees, in case it is "rad" then radians.}

\item{value}{the vector of rotation. Passed as the angle argument of rotate().}
}
\description{
Extracting the grid orientation

Extracting the grid orientation

Setting the orientation of a trigrid or hexagrid object

Setting the orientation of a trigrid or hexagrid object
}

