% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-attributes.R
\name{orientation}
\alias{orientation}
\alias{orientation,trigrid-method}
\alias{orientation<-}
\alias{orientation<-,trigrid-method}
\title{Extracting and setting the grid orientation}
\usage{
orientation(x, ...)

\S4method{orientation}{trigrid}(x, display = "deg", ...)

orientation(x) <- value

\S4method{orientation}{trigrid}(x) <- value
}
\arguments{
\item{x}{(\code{\link{trigrid}} or \code{\link{hexagrid}}): Input grid.}

\item{...}{Values passed on to the \code{\link[icosa]{rotate}} function.}

\item{display}{(\code{character}) The output unit. In case it is set to \code{"deg"} the output will be in degrees, in case it is \code{"rad"}, then radians.}

\item{value}{(\code{numeric}) The vector of rotation. Passed as the \code{angles} argument of \code{\link[icosa]{rotate}}.}
}
\value{
In case the function returns does, it returns the orientation angles of the grid (as \code{numeric}).
}
\description{
Extracting and setting the grid orientation
}
