% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\docType{methods}
\name{rotate,trigrid-method}
\alias{rotate,}
\alias{rotate,trigrid-method}
\alias{trigrid-rotate-method}
\title{Rotation method of trigrid and hexagrid objects}
\usage{
\S4method{rotate}{trigrid}(x, angles = "random", pivot = NA)
}
\arguments{
\item{x}{trigrid or hexagrid class object.}

\item{angles}{The vector of rotation (length=3) in radians. If set to "random", the rotation will be random (default).}

\item{pivot}{The pivot point of the rotation, numeric vector of xyz coordinates. Defaults to NA indicating that the rotation will be around the center of the grid.}
}
\value{
Another trigrid or hexagrid class object.
}
\description{
Rotation method of trigrid and hexagrid objects
}

