% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gridlayer-attributes.R
\docType{data}
\name{values}
\alias{values}
\alias{values,gridlayer-method}
\alias{values<-}
\alias{values<-,gridlayer-method}
\title{Extract and replace values from a gridlayer-derived object (e.g. \code{link{facelayer}}).}
\format{An object of class \code{standardGeneric} of length 1.}
\usage{
values(x,...)

\S4method{values}{gridlayer}(x)

values(x) <- value

\S4method{values}{gridlayer}(x) <- value
}
\arguments{
\item{x}{(\code{\link{facelayer}}) Object.}

\item{value}{(\code{logical}, \code{character} or \code{numeric}) Replacement values.}

\item{...}{Arguments passed to class-specific methods. (Not used.)}
}
\description{
The function will get the \code{@values} slot of a \code{\link{facelayer}} object.
}
\keyword{datasets}
